do not use container for long encrypted messages, prepare to multiple wire prefixes
Long encrypted messages are embed in containers and sent in multiple
splitted messages. Splitted messages aren't displayed and Silence stores
them to wait for the full container. This way to send encrypted messages
creates a risk to consider an unencrypted message as a contained one,
resulting in lost of unencrypted messages in rare (but plausible)
situations.

This commit changes the way to send encrypted messages. They aren't
splitted in containers anymore. Non contained encrypted messages will be
fully recognized and decrypted by previous Silence releases.

This commit do not change the way to receive and decrypt messages yet.
So it can still lost unencrypted messages. The way to receive messages
will be updated to do not take care of containers in the future, after a
1-month time frame. This is because once containers will be ignored,
contained encrypted messages will result in garbage. With a 1-month time
frame, most of users should have received the updated Silence release
that do not use containers, then allowing us to do not expect to receive
contained encrypted messages and not to lower user experience.

Also, this commit adds a function to update the type of a message. This
prepares the consequence of not losing messages: encrypted messages are
prefixed according to the expected type of the message, but this prefix
can be valid for multiple types.